+++ src/hooks/useCartSelection.ts
@@ -1,0 +1,50 @@
+import { useState, useEffect } from 'react';
+import { CartProduct } from '../types/cart';
+
+interface CartItemsResponse {
+  content: CartProduct[];
+}
+
+export const useCartSelection = (cartItems: CartItemsResponse | undefined) => {
+  const [checkedItems, setCheckedItems] = useState<number[]>([]);
+
+  useEffect(() => {
+    if (cartItems?.content && cartItems.content.length > 0) {
+      const itemIds = cartItems.content.map((item: CartProduct) => item.id);
+      setCheckedItems(itemIds);
+    } else {
+      setCheckedItems([]);
+    }
+  }, [cartItems]);
+
+  const isAllChecked = Boolean(
+    cartItems?.content &&
+      cartItems.content.length > 0 &&
+      checkedItems.length === cartItems.content.length,
+  );
+
+  const handleAllCheck = (checked: boolean) => {
+    if (!cartItems?.content) return;
+
+    if (checked) {
+      setCheckedItems([]);
+    } else {
+      const allIds = cartItems.content.map((item: CartProduct) => item.id);
+      setCheckedItems(allIds);
+    }
+  };
+
+  const toggleItem = (id: number) => {
+    setCheckedItems((prev) =>
+      prev.includes(id) ? prev.filter((itemId) => itemId !== id) : [...prev, id],
+    );
+  };
+
+  return {
+    checkedItems,
+    setCheckedItems,
+    isAllChecked,
+    handleAllCheck,
+    toggleItem,
+  };
+};
